requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0

# CORS Support
django-cors-headers==4.3.1

# JWT Authentication
djangorestframework-simplejwt==5.5.1

# Filtering
django-filter==24.1

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# Database
dj-database-url==2.1.0  # Database URL parsing for production
psycopg2-binary==2.9.9  # PostgreSQL adapter for Railway/Render
# mysqlclient==2.2.0  # MySQL adapter (optional - uncomment if using MySQL)

# Production Server
gunicorn==21.2.0  # WSGI HTTP Server for production
whitenoise==6.6.0  # Static file serving

# Additional Utilities
sqlparse==0.4.4
tzdata==2023.3
